implemented the new route and new controller for all cart functionality
