+++ client/app.js
@@ -3,6 +3,7 @@ import userRouter from "./routes/user.route.js"
 import cookieParser from "cookie-parser"
 import otpRouter from "./routes/otp.routes.js"
 import sellerRouter from "./routes/seller.route.js"
+import cartRouter from "./routes/cart.route.js"
 
 const app = express()
 
@@ -12,6 +13,7 @@ app.use(cookieParser())
 app.use("/api/user", userRouter)
 app.use("/api/otp", otpRouter)
 app.use("/api/seller", sellerRouter)
+app.use("/api/cart", cartRouter)
 
 //! for error
 app.use((err, req, res, next)=>{
